third_party/WebKit/LayoutTests/http/tests/inspector/resource-tree/resource-tree-events.html

<html>
<head>
<script src="../inspector-test.js"></script>
<script src="../resources-test.js"></script>
<script src="resource-tree-test.js"></script>

<script>

function test()
{
    InspectorTest.runAfterResourcesAreFinished(["inspector-test.js", "resources-test.js", "resource-tree-test.js"], step1);

    function step1()
    {
        // Reset resourceTreeModel.
        InspectorTest.resourceTreeModel.mainFrame._remove();

        for (var eventName in SDK.ResourceTreeModel.Events)
            InspectorTest.resourceTreeModel.addEventListener(SDK.ResourceTreeModel.Events[eventName], eventHandler.bind(this, eventName));
        for (var eventName in SDK.SecurityOriginManager.Events)
            InspectorTest.securityOriginManager.addEventListener(SDK.SecurityOriginManager.Events[eventName], eventHandler.bind(this, eventName));

        function eventHandler(eventName, event)
        {
            switch (eventName) {
            case "FrameAdded":
            case "FrameDetached":
                var frame = event.data;
                InspectorTest.addResult(`    ${eventName} : ${frame.id}`);
                break;
            case "FrameNavigated":
            case "MainFrameNavigated":
                var frame = event.data;
                InspectorTest.addResult(`    ${eventName} : ${frame.id} : ${frame.loaderId}`);
                break;
            case "SecurityOriginAdded":
            case "SecurityOriginRemoved":
                var securityOrigin = event.data;
                InspectorTest.addResult(`    ${eventName} : ${securityOrigin}`);
                break;
            default:
            }

        }

        // Simulate navigation to new render view: do not attach root frame.
        InspectorTest.addResult("Navigating root frame");
        InspectorTest.resourceTreeModel._frameNavigated(createFramePayload("root1"));
        InspectorTest.addResult("Navigating child frame 1");
        InspectorTest.resourceTreeModel._frameAttached("child1", "root1");
        InspectorTest.resourceTreeModel._frameNavigated(createFramePayload("child1", "root1"));
        InspectorTest.addResult("Navigating child frame 1 to a different URL");
        InspectorTest.resourceTreeModel._frameNavigated(createFramePayload("child1", "root1", "child1-new"));
        InspectorTest.addResult("Navigating child frame 2");
        InspectorTest.resourceTreeModel._frameAttached("child2", "root1");
        InspectorTest.resourceTreeModel._frameNavigated(createFramePayload("child2", "root1"));
        InspectorTest.addResult("Detaching child frame 1");
        InspectorTest.resourceTreeModel._frameDetached("child1");

        InspectorTest.addResult("Navigating root frame");
        InspectorTest.resourceTreeModel._frameAttached("root2");
        InspectorTest.resourceTreeModel._frameNavigated(createFramePayload("root2"));
        InspectorTest.completeTest();
    }

    function createFramePayload(id, parentId, name)
    {
        var framePayload = {};
        framePayload.id = id;
        framePayload.parentId = parentId || "";
        framePayload.loaderId = "loader-" + id;
        framePayload.name = "frame-" + id;
        framePayload.url = "http://frame/" + (name || id) + ".html";
        framePayload.securityOrigin = framePayload.url;
        framePayload.mimeType = "text/html";
        return framePayload;
    }
}

</script>

</head>

<body onload="runTest()">
<p>
Tests top frame navigation events.
</p>
<br>
</body>
</html>
